Refactor and added tests
