+++ core/src/test/scala/codesearch/core/regex/lexer/TokenizerSpec.scala
@@ -73,7 +73,7 @@ class TokenizerSpec extends FreeSpec with Matchers {
       )
     }
 
-    "Empty regex" - {
+    "Empty lexer" - {
       testParseAndRender("", Seq())
     }
